+++ src/runtime/components/Separator.vue
@@ -40,7 +40,7 @@ export interface SeparatorProps extends Pick<_SeparatorProps, 'decorative'> {
    * The orientation of the separator.
    * @defaultValue 'horizontal'
    */
-  orientation?: _SeparatorProps['orientation']
+  orientation?: Separator['variants']['orientation']
   class?: any
   b24ui?: Separator['slots']
 }