+++ test/components/Timeline.spec.ts
@@ -45,6 +45,7 @@ describe('Timeline', () => {
     ['with items', { props }],
     ['with modelValue', { props: { ...props, modelValue: 'design' } }],
     ['with defaultValue', { props: { ...props, defaultValue: 'design' } }],
+    ['with valueKey', { props: { ...props, valueKey: 'title', defaultValue: 'Design Phase' } }],
     ['with air-primary color', { props: { ...props, color: 'air-primary' } }],
     ...sizes.map((size: string) => [`with size ${size} horizontal`, { props: { ...props, size } }]),
     ...sizes.map((size: string) => [`with size ${size} vertical`, { props: { ...props, size, orientation: 'vertical' } }]),